# Conflicts:
#	src/algorithms/tracking/gnuradio_blocks/gps_l1_ca_dll_pll_tracking_cc.cc
#	src/algorithms/tracking/libs/tracking_discriminators.cc

+++ src/algorithms/tracking/gnuradio_blocks/galileo_e1_dll_pll_veml_tracking_cc.h
@@ -126,8 +126,8 @@ private:
     long d_if_freq;
     long d_fs_in;
 
-    double d_early_late_spc_chips;
-    double d_very_early_late_spc_chips;
+    float d_early_late_spc_chips;
+    float d_very_early_late_spc_chips;
 
     gr_complex* d_ca_code;
 
@@ -146,22 +146,22 @@ private:
 
     // remaining code phase and carrier phase between tracking loops
     double d_rem_code_phase_samples;
-    double d_rem_carr_phase_rad;
+    float d_rem_carr_phase_rad;
 
     // PLL and DLL filter library
     Tracking_2nd_DLL_filter d_code_loop_filter;
     Tracking_2nd_PLL_filter d_carrier_loop_filter;
 
     // acquisition
-    double d_acq_code_phase_samples;
-    double d_acq_carrier_doppler_hz;
+    float d_acq_code_phase_samples;
+    float d_acq_carrier_doppler_hz;
 
     // correlator
     Correlator d_correlator;
 
     // tracking vars
     double d_code_freq_chips;
-    double d_carrier_doppler_hz;
+    float d_carrier_doppler_hz;
     double d_acc_carrier_phase_rad;
     double d_acc_code_phase_secs;
 
@@ -175,9 +175,9 @@ private:
     // CN0 estimation and lock detector
     int d_cn0_estimation_counter;
     gr_complex* d_Prompt_buffer;
-    double d_carrier_lock_test;
-    double d_CN0_SNV_dB_Hz;
-    double d_carrier_lock_threshold;
+    float d_carrier_lock_test;
+    float d_CN0_SNV_dB_Hz;
+    float d_carrier_lock_threshold;
     int d_carrier_lock_fail_counter;
 
     // control vars
